[core] Inline RenderItem and RenderData
